feat: 총점 계산 로직 추가, totalScore type Boolean으로 수정

+++ src/main/java/com/gotcha/server/applicant/dto/response/CompletedApplicantsResponse.java
@@ -19,13 +19,13 @@ public class CompletedApplicantsResponse {
     private List<String> interviewers;
     private String email;
     private List<KeywordResponse> keywords;
-    private Integer totalScore;
+    private Double totalScore;
     private Integer ranking;
     private InterviewStatus interviewStatus;
     private List<OneLinerResponse> oneLiners;
 
     @Builder
-    public CompletedApplicantsResponse(String interviewName, List<String> interviewers, List<KeywordResponse> keywords, String email, LocalDate date, String name, InterviewStatus interviewStatus, Integer totalScore, Integer ranking, List<OneLinerResponse> oneLiners) {
+    public CompletedApplicantsResponse(String interviewName, List<String> interviewers, List<KeywordResponse> keywords, String email, LocalDate date, String name, InterviewStatus interviewStatus, Double totalScore, Integer ranking, List<OneLinerResponse> oneLiners) {
         this.interviewName = interviewName;
         this.interviewers = interviewers;
         this.keywords = keywords;
